requirements.txt
django==5.1 
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
django-environ==0.11.2
psycopg2==2.9.10
psycopg-pool==3.2.3
redis==5.2.0
pytest==8.3.3
pytest-django==4.9.0
faker==33.0.0
django-import-export==4.2.1
